Update calc method
Former-commit-id: 5093f5b46d0d13d07344bdc66ddc6d45284a367f

+++ src/main/java/com/tw/jiracalc/beans/CycleTimeBean.java
@@ -3,13 +3,13 @@ package com.tw.jiracalc.beans;
 import java.util.Map;
 
 public class CycleTimeBean {
-    private Map<String, Float> cycleTime;
+    private Map<String, Double> cycleTime;
 
-    public Map<String, Float> getCycleTime() {
+    public Map<String, Double> getCycleTime() {
         return cycleTime;
     }
 
-    public void setCycleTime(Map<String, Float> cycleTime) {
+    public void setCycleTime(Map<String, Double> cycleTime) {
         this.cycleTime = cycleTime;
     }
 }
